Goal.PercentComplete is working correctly for PercentGoalTasks.

+++ Code/CygSoft.SmartSession.Domain/Sessions/PercentSessionResult.cs
@@ -6,7 +6,7 @@ namespace CygSoft.SmartSession.Domain.Sessions
     {
         public double PercentCompleted { get; private set; }
 
-        public PercentSessionResult(DateTime dateTime, DateTime endTime, int percent) : base(dateTime, endTime)
+        public PercentSessionResult(DateTime dateTime, DateTime endTime, double percent) : base(dateTime, endTime)
         {
             if (percent < 0)
                 throw new ArgumentOutOfRangeException("Percent cannot be a negative value.");
